+++ pyslang/changelog.md
@@ -1,6 +1,9 @@
 # Change Log
 
-## [1.1.0] - 13/12/2025
+## [1.1.1] - 23/12/2025
+
+### Added
+- Keyword `match` for PyScript 1.8.0+
 
 ### Fixed
-- String highlight fixes.
+- Match regex for integer.